simulator_app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
certifi==2023.11.17
requests==2.31.0
python-dotenv==1.0.0